+++ tests/Fixtures/IndexableTestModel.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Swis\Laravel\Fulltext\Tests\Fixtures;
+namespace Saber13812002\Laravel\Fulltext\Tests\Fixtures;
 
-use Swis\Laravel\Fulltext\Indexable;
+use Saber13812002\Laravel\Fulltext\Indexable;
 
 class IndexableTestModel extends TestModel
 {